requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
pyaml==23.7.0
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
urllib3==2.0.4
xmltodict==0.13.0
